+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListView.swift
@@ -203,7 +203,7 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
                             }
                         }
                         withAnimation {
-                            scrollView.scrollTo(scrolledId, anchor: .bottom)
+                            scrollView.scrollTo(scrolledId, anchor: messageListConfig.scrollingAnchor)
                         }
                     }
                 }