+++ CHANGELOG.md
@@ -3,7 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### ✅ Added
+- Config to change the scrolling anchor (top/bottom) on messages
 
 # [4.25.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.25.0)
 _December 16, 2022_